en fin de transac, les prod acheté est stocké dans vente BDD

+++ goLikeBDD.php
@@ -7,27 +7,6 @@ $mode = $qq[0]; $idboug = intval($qq[1]);$idbeat = intval($qq[2]);
 
 $ok = true;
 
-//*** verif mode
-if($mode != 'likelike' && $mode != "dislikedislike"){  
-    echo "#modeexistepas#";
-    $ok = false;
-
-}
-
-//*** Verification du boug
-$req = $BDD->prepare("SELECT user_pseudo 
-                            FROM user
-                            WHERE user_id = ?");
-$req->execute(array($idboug));
-$verif_bg = $req->fetch();
-
-if(!isset($verif_bg['user_pseudo'])){  
-    echo "#bougexistepas#";
-    $ok = false;
-
-}
-
-
 //*** Verification du beat
 $req = $BDD->prepare("SELECT beat_title 
                             FROM beat
@@ -35,57 +14,76 @@ $req = $BDD->prepare("SELECT beat_title
 $req->execute(array($idbeat));
 $verif_b = $req->fetch();
 
-if(!isset($verif_b['beat_title'])){  
-    echo "#beatexistepas#";
-    $ok = false;
+if(isset($verif_b['beat_title'])){
+    //*** verif mode
+    if(($mode == 'likelike' || $mode == "dislikedislike") ) {  
 
-}
+        //*** Verification du boug
+        $req = $BDD->prepare("SELECT user_pseudo 
+                            FROM user
+                            WHERE user_id = ?");
+        $req->execute(array($idboug));
+        $verif_bg = $req->fetch();
+
+        if(!isset($verif_bg['user_pseudo'])){  
+            echo "#bougexistepas#";
+            $ok = false;
+
+        }
 
 
-if($mode == "likelike") {
-{ // ensuite on verifie si ce beat est deja like
-    $req = $BDD->prepare("SELECT id
+        if($mode == "likelike") {
+            // ensuite on verifie si ce beat est deja like
+            $req = $BDD->prepare("SELECT id
                             FROM likelike
                             WHERE like_user_id = ? AND like_beat_id = ?
                                 ");
-    $req->execute(array($idboug,$idbeat));
-    $p = $req->fetch();
+            $req->execute(array($idboug,$idbeat));
+            $p = $req->fetch();
 
-    if(isset($p['id'])){
-        $ok = false;
-        echo "Cette beat est deja liké !";
-    }
-}
+            if(isset($p['id'])){
+                $ok = false;
+                echo "Cette beat est deja liké !";
+            }
 
-    if($ok){
 
-        $req = $BDD->prepare("INSERT INTO likelike (like_beat_id,like_user_id) VALUES (?, ?)"); 
-        $req->execute(array($idbeat,$idboug));
+            if($ok){
 
-        echo "*** like ***";
-    }
-}else if ($mode == "dislikedislike") {
-    { // ensuite on verifie si ce beat est deja like
-    $req = $BDD->prepare("SELECT id
+                $req = $BDD->prepare("INSERT INTO likelike (like_beat_id,like_user_id) VALUES (?, ?)"); 
+                $req->execute(array($idbeat,$idboug));
+
+                echo "*** like ***";
+            }
+            
+        }else if ($mode == "dislikedislike") {
+            { // ensuite on verifie si ce beat est deja like
+                $req = $BDD->prepare("SELECT id
                             FROM likelike
                             WHERE like_user_id = ? AND like_beat_id = ?
                                 ");
-    $req->execute(array($idboug,$idbeat));
-    $p = $req->fetch();
+                $req->execute(array($idboug,$idbeat));
+                $p = $req->fetch();
 
-    if(!isset($p['id'])){
-        $ok = false;
-        echo "Cette beat n'est pas liké donc tu veux quoi ?";
-    }
-}
+                if(!isset($p['id'])){
+                    $ok = false;
+                    echo "Cette beat n'est pas liké donc tu veux quoi ?";
+                }
+            }
 
-    if($ok){
+            if($ok){
 
-        $req = $BDD->prepare("DELETE FROM likelike 
+                $req = $BDD->prepare("DELETE FROM likelike 
  WHERE like_user_id = ? AND like_beat_id = ?"); 
-        $req->execute(array($idboug,$idbeat));
+                $req->execute(array($idboug,$idbeat));
 
-        echo "*** dislike ***";
+                echo "*** dislike ***";
+            }
+        }
+
+    } else{
+        echo "#modeexistepas#";
     }
+} else {
+    echo "#beatexistepas#";
 }
 ?>
